Add files via upload
Input01 sheet as Excel input to be provided by the user and read by the program as input, PO Number column is the number generated or written by the program back to Excel. Output01 and Output02 would be the template of the invoice written by the program and converted to PDF depending on the number of records process from the Input01 rows. The maximum number of entry lines or Descriptions will be 5 rows per invoice template.